The first run of the prototype for Mia
The purpose of this prototype is to test how player choice can effect game environments.
The player will be asked to create a newspaper story with various sentences provided through “research” to the reporter. The end result will calculate the sway of the news created by the player. A dialogue will appear at the end that highlights the changes to the world that occurred because of the news choices of the player.